refactor: rename feed to fetch
