+++ .mockery.yaml
@@ -7,7 +7,7 @@ mockname: "Mock{{.InterfaceName}}"
 outpkg: "{{.PackageName}}"
 
 packages:
-  github.com/ericbutera/amalgam/data-pipeline/temporal/feed/internal/transforms:
+  github.com/ericbutera/amalgam/data-pipeline/temporal/feed_fetch/internal/transforms:
     interfaces:
       Transforms:
   github.com/ericbutera/amalgam/data-pipeline/temporal/internal/bucket:
