+++ data-pipeline/temporal/feed_fetch/worker/main.go
@@ -5,9 +5,9 @@ import (
 	"log/slog"
 	"os"
 
-	app "github.com/ericbutera/amalgam/data-pipeline/temporal/feed"
-	"github.com/ericbutera/amalgam/data-pipeline/temporal/feed/internal/config"
-	"github.com/ericbutera/amalgam/data-pipeline/temporal/feed/internal/transforms"
+	app "github.com/ericbutera/amalgam/data-pipeline/temporal/feed_fetch"
+	"github.com/ericbutera/amalgam/data-pipeline/temporal/feed_fetch/internal/config"
+	"github.com/ericbutera/amalgam/data-pipeline/temporal/feed_fetch/internal/transforms"
 	"github.com/ericbutera/amalgam/data-pipeline/temporal/internal/bucket"
 	"github.com/ericbutera/amalgam/data-pipeline/temporal/internal/client"
 	"github.com/ericbutera/amalgam/data-pipeline/temporal/internal/feeds"
